Modify Panda.java
Modify Panda Class so that reply Method takes generic input instead, so
that it can take String, as well as Objects of TaskList Class and Task
Class as well.

To account for more actions from command line, switch will be used
instead of else-if.

Certain Methods in Panda.java are not defined yet, and TaskList class
will be modified afterwards to define those Methods to be used.