Modify code to use | as separator for String split
The initial code used \u2022 as the separator, in hopes of having
a separator that is not available on the keyboard, so that task
description will likely not contain it.

However, upon running with javaFX GUI, there appears to be issues with
how it is parsed and read, affecting the splitting of the String.

Hence, | will be used instead so that there will not be any more issues
with the String not being split.