Add HelpAction.java
For users who do not know how to use the program, they can type "help"
to get the list of actions that can be used.

The other classes are also updated, including prompts to inform user
of the "help" action.

This will help users who are new, as well as those who have not used
the program in a long time.